BS-68 | Siva Reddy | Updated Rest End Point for Diagnosis Search
