+++ bahmnicore-omod/src/main/resources/config.xml
@@ -33,11 +33,8 @@
         <require_module>org.openmrs.module.episodes</require_module>
         <require_module>org.openmrs.module.auditlog</require_module>
         <require_module>org.bahmni.module.bahmnicommons</require_module>
+        <require_module>org.bahmni.module.terminologyservices</require_module>
     </require_modules>
-
-    <aware_of_modules>
-        <aware_of_module>org.bahmni.module.terminologyservices</aware_of_module>
-    </aware_of_modules>
     <!-- Extensions -->
 
     <!-- AOP -->
